+++ ArmPkg/Drivers/ArmGic/GicV3/ArmGicV3Dxe.c
@@ -423,7 +423,7 @@ GicV3DxeInitialize (
       for (Index = 8; Index < (mGicNumInterrupts / 4); Index++) {
         MmioWrite32 (
           mGicDistributorBase + ARM_GIC_ICDIPTR + (Index * 4),
-          CpuTarget
+          (UINT32)CpuTarget      // MU_CHANGE - ARM64 VS change
           );
       }
     }
@@ -460,7 +460,7 @@ GicV3DxeInitialize (
     for (Index = 0; Index < (mGicNumInterrupts - 32); Index++) {
       MmioWrite64 (
         mGicDistributorBase + ARM_GICD_IROUTER + (Index * 8),
-        CpuTarget
+        (UINT32)CpuTarget   // MU_CHANGE - ARM64 VS change
         );
     }
   }
